+++ lib/social_scribe/google_calendar.ex
@@ -1,37 +1,3 @@
-defmodule SocialScribe.GoogleCalendarClient do
-  @base_url "https://www.googleapis.com/calendar/v3"
-
-  # TODO: Mock for testing
-  def list_events(token, start_time, end_time, calendar_id \\ "primary") do
-    Tesla.get(client(token), "/calendars/#{calendar_id}/events",
-      query: [
-        timeMin: Timex.format!(start_time, "{RFC3339}"),
-        timeMax: Timex.format!(end_time, "{RFC3339}"),
-        singleEvents: true,
-        orderBy: "startTime"
-      ]
-    )
-    |> case do
-      {:ok, %Tesla.Env{status: 200, body: body}} ->
-        {:ok, body}
-
-      {:ok, %Tesla.Env{status: status, body: error_body}} ->
-        {:error, {status, error_body}}
-
-      {:error, reason} ->
-        {:error, reason}
-    end
-  end
-
-  defp client(token) do
-    Tesla.client([
-      {Tesla.Middleware.BaseUrl, @base_url},
-      {Tesla.Middleware.Headers, [{"Authorization", "Bearer #{token}"}]},
-      Tesla.Middleware.JSON
-    ])
-  end
-end
-
 defmodule SocialScribe.GoogleCalendar do
   @moduledoc """
   Fetches and syncs Google Calendar events.
@@ -39,7 +5,7 @@ defmodule SocialScribe.GoogleCalendar do
 
   require Logger
 
-  alias SocialScribe.GoogleCalendarClient
+  alias SocialScribe.GoogleCalendar.Client
   alias SocialScribe.Accounts
   alias SocialScribe.Accounts.UserCredential
   alias SocialScribe.Calendar
@@ -64,7 +30,7 @@ defmodule SocialScribe.GoogleCalendar do
   defp fetch_and_sync_for_credential(%UserCredential{} = credential) do
     with {:ok, token} <- ensure_valid_token(credential),
          {:ok, %{"items" => items}} <-
-           GoogleCalendarClient.list_events(
+           Client.list_events(
              token,
              DateTime.utc_now() |> Timex.beginning_of_day() |> Timex.shift(days: -1),
              DateTime.utc_now() |> Timex.end_of_day() |> Timex.shift(days: 7)
